encoders/lidarclip/requirements.txt
einops==0.8.1
ipdb==0.13.13
lyft_dataset_sdk==0.0.8
matplotlib==3.5.2
mmcv_full==1.3.9
mmdet==2.14.0
mmengine==0.10.6
mmsegmentation==0.14.1
numba==0.48.0
numpy==1.19.5
numpy==1.20.2
nuscenes_devkit==1.1.10
open3d==0.17.0
openai_clip==1.0.1
opencv_python==4.11.0.86
pandas==1.3.5
Pillow==9.5.0
Pillow==8.2.0
Pillow==12.0.0
pycocotools==2.0.7
pyquaternion==0.9.9
pytest==7.4.4
pytorch_lightning==1.6.2
scikit_image==0.19.3
scipy==1.7.3
seaborn==0.13.2
setuptools==52.0.0.post20210125
Shapely==1.8.5
tensorflow==2.20.0
terminaltables==3.1.10
torch==1.9.0
tqdm==4.65.2
tqdm==4.51.0
trimesh==2.35.39
waymo_open_dataset==1.0.1
